More PIDs & Gyroscope
Continued working on configuring the PIDs, going well, just require elevator. Added a Pigeon2 (a gyroscope) to help us prevent tipping by lowering the elevator by a certain amount once passing a threshold of a certain degrees (this is in the constants file). Very happy and worked great :D - Ethan