Exact Align working
It works now. Changed the yaw tolerance, but also made it so that the yaw gets centralized to 0 is straight instead of 180. also made it so that the yaw needs to go a bit farther towards the goal before stopping (to prevent oscillation due to error in photon). One final change is that it has to not be able to see the target for more than 2 frames (to prevent random stopping midway due to photon not being able to see the tag from movement)

-Ethan & Juli